add geometry.hpp geometry.cpp
