+++ projects/test/test/test.vcxproj
@@ -151,12 +151,14 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\sources\include\color.cpp" />
+    <ClCompile Include="..\..\..\sources\include\geometry.cpp" />
     <ClCompile Include="..\..\..\sources\include\piece.cpp" />
     <ClCompile Include="..\..\..\sources\include\utility.cpp" />
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\sources\include\color.hpp" />
+    <ClInclude Include="..\..\..\sources\include\geometry.hpp" />
     <ClInclude Include="..\..\..\sources\include\piece.hpp" />
     <ClInclude Include="..\..\..\sources\include\utility.hpp" />
   </ItemGroup>
