+++ sources/include/geometry.hpp
@@ -1,0 +1,10 @@
+#ifndef _GEOMETRY_HPP_
+#define _GEOMETRY_HPP_
+
+#include"utility.hpp"
+#include"piece.hpp"
+#include<vector>
+
+Point getPutPoint(std::vector<Piece>&, std::vector<putData>&,std::vector<std::vector<Point>>&);
+
+#endif
